# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ReturnOps API
  version: 1.0.0
  description: Versioned API contract for the ReturnOps clean-room portfolio demo.
    All business data is fictional and isolated by browser visitor.
paths:
  /api/v1/demo/reset/:
    post:
      operationId: resetDemoDataset
      description: Atomically replaces only the current visitor's data with the canonical
        fictional scenario and refreshes its expiration.
      summary: Reset the current visitor's fictional dataset
      tags:
      - Demo session
      security:
      - visitorSession: []
        csrfToken: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DemoResetResponse'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
  /api/v1/operations/returns/:
    get:
      operationId: listOperationsReturns
      summary: List non-draft returns for operational review
      parameters:
      - in: query
        name: ordering
        schema:
          enum:
          - -updated_at
          - updated_at
          - -total_value
          - total_value
          - reference
          type: string
          default: -updated_at
          minLength: 1
        description: |-
          * `-updated_at` - -updated_at
          * `updated_at` - updated_at
          * `-total_value` - -total_value
          * `total_value` - total_value
          * `reference` - reference
      - in: query
        name: page
        schema:
          type: integer
          minimum: 1
          default: 1
      - in: query
        name: page_size
        schema:
          type: integer
          maximum: 50
          minimum: 1
      - in: query
        name: search
        schema:
          type: string
          default: ''
          maxLength: 120
      - in: query
        name: status
        schema:
          enum:
          - SUBMITTED
          - NEEDS_INFORMATION
          - APPROVED
          - REJECTED
          - null
          type: string
          nullable: true
          minLength: 1
        description: |-
          * `SUBMITTED` - SUBMITTED
          * `NEEDS_INFORMATION` - NEEDS_INFORMATION
          * `APPROVED` - APPROVED
          * `REJECTED` - REJECTED
      tags:
      - Operations queue
      security:
      - visitorSession: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedReturnRequestSummaryList'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
  /api/v1/operations/returns/{id}/:
    get:
      operationId: retrieveOperationsReturn
      summary: Retrieve the expandable operations detail
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: Opaque return request UUID scoped to the current visitor.
        required: true
      tags:
      - Operations queue
      security:
      - visitorSession: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnRequestDetail'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
  /api/v1/operations/returns/{id}/transition/:
    post:
      operationId: transitionOperationsReturn
      description: Transitions SUBMITTED to NEEDS_INFORMATION, APPROVED, or REJECTED.
        Notes are mandatory when requesting information or rejecting.
      summary: Apply an operations state transition
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: Opaque return request UUID scoped to the current visitor.
        required: true
      tags:
      - Operations queue
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperationsTransitionRequest'
            examples:
              RequestMoreInformation:
                value:
                  target_status: NEEDS_INFORMATION
                  note: Add a fictional serial-number image.
                summary: Request more information
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/OperationsTransitionRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/OperationsTransitionRequest'
        required: true
      security:
      - visitorSession: []
        csrfToken: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnRequestDetail'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
  /api/v1/returns/:
    get:
      operationId: listCustomerReturns
      summary: List returns in the current visitor sandbox
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      tags:
      - Customer returns
      security:
      - visitorSession: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedReturnRequestSummaryList'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
    post:
      operationId: createCustomerReturn
      summary: Create an empty draft return
      tags:
      - Customer returns
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReturnRequestWriteRequest'
            examples:
              NewFictionalReturn:
                value:
                  order_reference: ORD-90001
                  customer_name: Taylor Example
                  customer_email: taylor@example.com
                summary: New fictional return
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ReturnRequestWriteRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ReturnRequestWriteRequest'
        required: true
      security:
      - visitorSession: []
        csrfToken: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnRequestDetail'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
  /api/v1/returns/{id}/:
    get:
      operationId: retrieveCustomerReturn
      summary: Retrieve a return with items, evidence, and timeline
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: Opaque return request UUID scoped to the current visitor.
        required: true
      tags:
      - Customer returns
      security:
      - visitorSession: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnRequestDetail'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
    patch:
      operationId: updateCustomerReturn
      summary: Update mutable fields on a draft or needs-information return
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: Opaque return request UUID scoped to the current visitor.
        required: true
      tags:
      - Customer returns
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedReturnRequestWriteRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedReturnRequestWriteRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedReturnRequestWriteRequest'
      security:
      - visitorSession: []
        csrfToken: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnRequestDetail'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
    delete:
      operationId: deleteCustomerDraft
      summary: Delete a draft return
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: Opaque return request UUID scoped to the current visitor.
        required: true
      tags:
      - Customer returns
      security:
      - visitorSession: []
        csrfToken: []
      responses:
        '204':
          description: No response body
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
  /api/v1/returns/{id}/items/:
    post:
      operationId: addCustomerReturnItem
      summary: Add an item to a mutable return
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: Opaque return request UUID scoped to the current visitor.
        required: true
      tags:
      - Customer returns
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReturnItemWriteRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ReturnItemWriteRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ReturnItemWriteRequest'
        required: true
      security:
      - visitorSession: []
        csrfToken: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnItem'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Stable API error envelope.
...
components:
  schemas:
    APIError:
      type: object
      description: Stable error envelope returned by every handled API exception.
      properties:
        code:
          type: string
          readOnly: true
        detail:
          type: string
          readOnly: true
        fields:
          readOnly: true
      required:
      - code
      - detail
      - fields
...
    DemoResetResponse:
      type: object
      properties:
        reset_at:
          type: string
          format: date-time
          readOnly: true
        return_count:
          type: integer
          readOnly: true
        message_code:
          type: string
          readOnly: true
      required:
      - message_code
      - reset_at
      - return_count
...
    OperationsTransitionRequest:
      type: object
      description: Reject unknown and server-controlled fields instead of ignoring
        them.
      properties:
        target_status:
          $ref: '#/components/schemas/TargetStatusEnum'
        note:
          type: string
          default: ''
          maxLength: 2000
      required:
      - target_status
    PaginatedReturnRequestSummaryList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/ReturnRequestSummary'
...
    ReturnRequestWriteRequest:
      type: object
      description: Reject unknown and server-controlled fields instead of ignoring
        them.
      properties:
        order_reference:
          type: string
          minLength: 1
          maxLength: 40
        customer_name:
          type: string
          minLength: 1
          maxLength: 120
        customer_email:
          type: string
          format: email
          minLength: 1
          maxLength: 254
      required:
      - customer_email
      - customer_name
      - order_reference
...
  securitySchemes:
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRFToken
      description: Required with the visitor session for POST, PATCH, and DELETE.
        Read its value from the non-HttpOnly returnops_csrftoken cookie.
    visitorSession:
      type: apiKey
      in: cookie
      name: returnops_sessionid
      description: Opaque HttpOnly Django session cookie. Obtain it from GET /api/v1/session/;
        visitor UUIDs are never accepted.
...